+++ src/app/app.component.html
@@ -2,9 +2,8 @@
 <h3> Click below button to publish message </h3>
 
 
-<button (click)="orderPlaced($event)">Publish Order Placed</button>
->
-<button (click)="orderPaid($event)">Publish Order Paid</button>
+<button (click)="orderPlaced($event)">Publish Order Placed</button> <br> <hr>
 
-<button (click)="orderShipped($event)">Publish Order Shipped</button
+<button (click)="orderPaid($event)">Publish Order Paid</button><br><hr>
 
+<button (click)="orderShipped($event)">Publish Order Shipped</button>
