+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { PubsubService } from '@fsms/angular-pubsub';
 import { OrderPaid } from './messages/order-paid-message';
 import { OrderShipped } from './messages/order-shipped-message';
@@ -9,16 +9,11 @@ import { OrderPlaced, OrderPlacedType } from './messages/placeorder-message';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
 })
-export class AppComponent {
+export class AppComponent implements OnInit {
   title = 'Pubsub Framework Example App';
 
-  constructor(private pubsubService: PubsubService) {
-    this.pubsubService.subscribe({
-      messageType: OrderPlacedType,
-      callback: (msg) => console.log('received', msg),
-    });
-  }
-
+  constructor(private pubsubService: PubsubService) {}
+  ngOnInit(): void {}
   orderPlaced($event: KeyboardEvent) {
     $event.preventDefault();
     this.pubsubService.publish(new OrderPlaced('20 Apples'));
