+++ src/app/app.module.ts
@@ -6,11 +6,13 @@ import {
   PrepareOrderService,
   ShipOrderService,
 } from './services/ship-order.service';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [
     BrowserModule,
+    FormsModule,
     PubSubModule.forRoot([ShipOrderService, PrepareOrderService]),
   ],
   providers: [],
